% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readecb-package.R
\docType{package}
\name{readecb-package}
\alias{readecb}
\alias{readecb-package}
\title{readecb: Access European Central Bank Data}
\description{
Provides clean, tidy access to statistical data published by the European Central Bank (ECB) via the ECB Data Portal API \url{https://data.ecb.europa.eu}. Covers policy interest rates, EURIBOR, euro exchange rates, harmonised consumer price inflation (HICP), euro area yield curves, the euro short-term rate (ESTR), monetary aggregates (M1, M2, M3), mortgage and lending rates, GDP, unemployment, and government debt-to-GDP. Each dataset has a dedicated function that abstracts away the underlying SDMX key structure, so users do not need to know series codes. A generic fetcher is also provided for direct access to any of the ECB's 100-plus dataflows. Data is downloaded on first use and cached locally for subsequent calls.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/charlescoverdale/readecb}
  \item Report bugs at \url{https://github.com/charlescoverdale/readecb/issues}
}

}
\author{
\strong{Maintainer}: Charles Coverdale \email{charlesfcoverdale@gmail.com}

}
\keyword{internal}
